Refactor engine tools and modularize commands
