+++ core/url/base64url.ts
@@ -1,0 +1,31 @@
+export function base64UrlEncode(bytes: Uint8Array) {
+  if (typeof Buffer !== 'undefined') {
+    return Buffer.from(bytes)
+      .toString('base64')
+      .replace(/=+$/g, '')
+      .replace(/\+/g, '-')
+      .replace(/\//g, '_');
+  }
+  let s = '';
+  for (let i = 0; i < bytes.length; i += 1) {
+    s += String.fromCharCode(bytes[i]);
+  }
+  const b64 = (typeof btoa === 'function' ? btoa(s) : '')
+    .replace(/=+$/g, '')
+    .replace(/\+/g, '-')
+    .replace(/\//g, '_');
+  return b64;
+}
+
+export function base64UrlDecodeToBytes(str: string) {
+  const b64 = str.replace(/-/g, '+').replace(/_/g, '/');
+  if (typeof Buffer !== 'undefined') {
+    return new Uint8Array(Buffer.from(b64, 'base64'));
+  }
+  const bin = atob(b64);
+  const out = new Uint8Array(bin.length);
+  for (let i = 0; i < bin.length; i += 1) {
+    out[i] = bin.charCodeAt(i);
+  }
+  return out;
+}
